Update requirements, add benchmarks
updating requirements: pyo3 and ndarray seem to be working on stable
rust. Unfortunately adding the benchmarks means we are reliant on the
unstable test feature, so we can't use stable yet.

Added benchmarks. Generates some random arrays with seeds for
reproducibility, and runs each step of maximin_tree_query.

Would like to use flamegraph, but cannot figure out how to get it to
work with a pyo3 library